.refactoring on MainLayout.kt - Extracted composable to own files: Splitcounter.kt and TotalTipAmount.kt - Created package for common composable, moved RoundTheTipSwitch.kt and EditNumber.kt to it.
