+++ app/src/main/java/com/exercise/matipv2/ui/MainLayout.kt
@@ -1,23 +1,13 @@
 package com.exercise.matipv2.ui
 
 import android.annotation.SuppressLint
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -25,19 +15,19 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.exercise.matipv2.R
-import com.exercise.matipv2.components.EditNumber
-import com.exercise.matipv2.components.RoundTheTipSwitch
+import com.exercise.matipv2.components.common.EditNumber
+import com.exercise.matipv2.components.common.RoundTheTipSwitch
+import com.exercise.matipv2.components.main.SplitCounter
+import com.exercise.matipv2.components.main.TotalTipAmount
 
 @SuppressLint("VisibleForTests")
 @Composable
@@ -55,6 +45,7 @@ fun MainLayout(viewModel: MainLayoutViewModel) {
 
         /* TODO("Add Branding") */
 
+        /* Title */
         Text(
             modifier = Modifier
                 .padding(bottom = dimensionResource(R.dimen.padding_mid))
@@ -64,7 +55,7 @@ fun MainLayout(viewModel: MainLayoutViewModel) {
             style = MaterialTheme.typography.titleLarge,
         )
 
-        /* Function to edit Bill Amount */
+        /* Edit Total Bill Amount */
         EditNumber(
             modifier = Modifier
                 .padding(bottom = dimensionResource(R.dimen.padding_mid))
@@ -79,7 +70,7 @@ fun MainLayout(viewModel: MainLayoutViewModel) {
             leadingIcon = R.drawable.attach_money
         )
 
-        /* Function to edit Tip percentage */
+        /* Edit Tip percentage */
         EditNumber(
             modifier = Modifier
                 .padding(bottom = 40.dp)
@@ -94,68 +85,13 @@ fun MainLayout(viewModel: MainLayoutViewModel) {
             leadingIcon = R.drawable.percent
         )
 
-        /* Split value */
-        Row(
-            modifier = Modifier
-                .padding(bottom = dimensionResource(R.dimen.padding_mid))
-                .fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text(
-                text = stringResource(R.string.split),
-                fontWeight = FontWeight.Bold,
-                style = MaterialTheme.typography.titleLarge,
-            )
-            Spacer(modifier = Modifier.padding(20.dp))
-            Button(
-                onClick = { viewModel.decreaseCounter() },
-                shape = CircleShape,
-                colors = ButtonDefaults.buttonColors()
-            ) { Text(text = "-") }
-            AnimatedContent(
-                targetState = uiState.splitShare,
-                transitionSpec = {
-                    if (targetState > initialState) {
-                        slideInVertically { -it } togetherWith slideOutVertically { it }
-                    } else {
-                        slideInVertically { it } togetherWith slideOutVertically { -it }
-                    }
-                }, label = "Animated Counter"
-            ) { count ->
-                Text(
-                    text = "$count",
-                    style = MaterialTheme.typography.displaySmall,
-                    textAlign = TextAlign.Center,
-                )
-            }
-            Button(
-                onClick = { viewModel.increaseCounter() },
-                shape = CircleShape,
-                colors = ButtonDefaults.buttonColors()
-            ) { Text(text = "+") }
-        }
+        SplitCounter(
+            splitUiState = uiState.splitShare,
+            onclickPositive = { viewModel.increaseCounter() },
+            onclickNegative = { viewModel.decreaseCounter() }
+        )
 
-        /* Text Box for total Tip Amount */
-        Row(
-            modifier = Modifier
-                .padding(bottom = dimensionResource(R.dimen.padding_sml))
-                .fillMaxSize(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text(
-                text = stringResource(R.string.tip_amount),
-                fontWeight = FontWeight.Bold,
-                style = MaterialTheme.typography.titleLarge,
-            )
-            Text(
-                modifier = Modifier.testTag("TipAmount"),
-                text = viewModel.finalTip(),
-                fontWeight = FontWeight.Bold,
-                style = MaterialTheme.typography.titleLarge,
-            )
-        }
+        TotalTipAmount(viewModel.finalTip())
 
         RoundTheTipSwitch(
             modifier = Modifier
